fix prisma build, remove tsup, fix backend imports
