+++ apps/backend/src/services/TelemetryService.ts
@@ -1,5 +1,6 @@
 import { PostHogSender } from "@codemod-com/telemetry";
-import type { TelemetryEvents } from "../telemetry";
+
+import type { TelemetryEvents } from "../telemetry.js";
 
 export const telemetryService = new PostHogSender<TelemetryEvents>({
   cloudRole: "",
